requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.5.3
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
cchardet==2.1.6
celery==4.4.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-crispy-forms==1.9.2
django-model-utils==4.0.0
django-tempus-dominus==5.1.2.13
django-utils==0.0.2
django-widget-tweaks==1.4.8
elasticsearch==7.8.1
fake-useragent==0.1.11
filelock==3.0.12
geographiclib==1.50
geopy==2.0.0
googletransx==2.4.2
gunicorn==20.0.4
idna==2.10
kombu==4.6.11
multidict==4.7.6
numpy==1.19.1
pandas==1.1.0
pipenv==2020.8.13
psycopg2==2.8.5
pycares==3.1.1
pycparser==2.20
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
twint @ git+https://github.com/twintproject/twint.git@d7506f5a67cd96a723b92776e2ad03ed6dfacaec
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
yarl==1.5.1
